requirements.txt
python-telegram-bot==21.7
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
ta-lib==0.4.28
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
redis==5.0.1
ccxt==4.2.23
yfinance==0.2.33
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.5.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2